requirements.txt
pytest
requests
playwright
pytest-progress
python-dotenv